// msfs-avionics-mirror/src/sdk/flightplan/FlightPlan.ts
export enum PlanChangeType {
  Added = 'Added',
  Inserted = 'Inserted',
  Removed = 'Removed',
  Changed = 'Changed',
  Cleared = 'Cleared'
}

export enum ActiveLegType {
  Lateral = 'Lateral',
  Vertical = 'Vertical',
  Calculating = 'Calculating'
}

export enum OriginDestChangeType {
  OriginAdded = 'OriginAdded',
  OriginRemoved = 'OriginRemoved',
  DestinationAdded = 'DestinationAdded',
  DestinationRemoved = 'DestinationRemoved'
}

/** Direct To Metadata for Flight Plan. */
export interface DirectToData {
  /** The segment index of the direct to target. */
  segmentIndex: number;
  /** The segment leg index of the direct to target. */
  segmentLegIndex: number;
}

/**
 * Events generated by changes in the flight plan.
 */
export interface PlanEvents {
  /**
   * An event generated when there is a change in a leg.
   * @param segmentIndex The index of the segment.
   * @param index The index of the leg.
   * @param type The type of change.
   * @param leg The leg that was changed.
   */
  onLegChanged?(segmentIndex: number, index: number, type: PlanChangeType, leg?: LegDefinition): void;

  /**
   * An event generated when there is a change in a segment.
   * @param index The index of the segment.
   * @param type The type of change.
   * @param segment The segment that was changed.
   */
  onSegmentChanged?(index: number, type: PlanChangeType, segment?: FlightPlanSegment): void;

  /**
   * An event generated when an active leg is changed.
   * @param index The index of the leg.
   * @param segmentIndex The index of the flight plan segment.
   * @param previousSegmentIndex The index of the previously active segment.
   * @param previousLegIndex The index of the previously active leg in the previously active segment.
   * @param type The type of leg that was changed.
   */
  onActiveLegChanged?(index: number, segmentIndex: number, legIndex: number, previousSegmentIndex: number, previousLegIndex: number, type: ActiveLegType): void;

  /**
   * An event fired when a flight plan vectorization completes.
   * @param index The index that the path was generated from.
   */
  onCalculated?(index?: number): void;

  /**
   * An event fired when the origin or destination changes.
   * @param type The origin/destination change type.
   * @param airport The airport that was changed.
   */
  onOriginDestChanged?(type: OriginDestChangeType, airport?: string): void;

  /**
   * An event fired when the procedure details change.
   * @param details The details that were changed.
   */
  onProcedureDetailsChanged?(details: ProcedureDetails): void;

  /**
   * An event fired when user data is set in the plan.
   * @param key The key of the user data.
   * @param data The data that was set.
   */
  onUserDataSet?(key: string, data: any): void;

  /**
   * An event fired when user data is deleted in the plan.
   * @param key The key of the user data to delete.
   */
  onUserDataDelete?(key: string): void;

  /**
   * An event fired when direct to data is changed in the plan.
   * @param directToData The direct to data.
   */
  onDirectDataChanged?(directToData: DirectToData): void;
}